Rename lib.py to utils.py and add test cases for natural sort
